Support base32 encoded public key subdomains
Subdomains can only have 63 characters, so public key can't be used.  Base64 is case sensitive and domains are not.  I used base32 which shortened the public key enough to fit in the subdomain, and all the characters are case insensitive and url safe.

This change required me to bundle the front-end in order to use the base32-encoding library.  I chose Svelte.