Support subdomains.  Refactor.  Migrate focus to integration website. (#4)
* Update ssl certificates

Use a more cohesive genration script
Use *.local.computer, a public service domain that supports wildcards and resolves to localhost.

* Support base32 encoded public key subdomains

Subdomains can only have 63 characters, so public key can't be used.  Base64 is case sensitive and domains are not.  I used base32 which shortened the public key enough to fit in the subdomain, and all the characters are case insensitive and url safe.

This change required me to bundle the front-end in order to use the base32-encoding library.  I chose Svelte.

* Add support for base32 public key subdomains

* Separate web ui repo.  Configure to be used in hyperdrive-gateway-integration

* Stream data from hyperdrive to response

* Refactor and cleanup.

* Small fixes.  Rewrite unit tests for refactors.

* Update README

* Rename to hyperdrive-gateway-server